ASR_FastApi/requirements.txt
fastapi
uvicorn
onnxruntime
numpy
soundfile
nemo_toolkit[asr]
torch
omegaconf
